feat:#470-Slug logica verplaatst binnen eigen class
